% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{add_cols_and_reorder}
\alias{add_cols_and_reorder}
\title{Add missing columns with names in `by` not in the column names of `tab`. Order
the table, ignoring the first column in the order given in the `by` vector.
Set all NAs to 0 in the data columns only.}
\usage{
add_cols_and_reorder(tab, by)
}
\arguments{
\item{tab}{data frame with years as the first column, and N columns of data
with names}

\item{by}{a vector of names, which correspond to the names of the columns in tab}
}
\value{
a modified data frame
}
\description{
Add missing columns with names in `by` not in the column names of `tab`. Order
the table, ignoring the first column in the order given in the `by` vector.
Set all NAs to 0 in the data columns only.
}
