% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{basic_table}
\alias{basic_table}
\title{General table for the herring assessment.}
\usage{
basic_table(tab, cap = "", translate = FALSE, ...)
}
\arguments{
\item{tab}{data.frame as read in by [readr::read_csv()]}

\item{cap}{caption for table}

\item{translate}{Logical. Translate to french if TRUE}

\item{...}{arguments passed to [csas_table()]}
}
\value{
a [csasdown::csas_table()]
}
\description{
General table for the herring assessment.
}
