% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{consecutive_group}
\alias{consecutive_group}
\title{Function to add a new column indicating group ID based on sequential data}
\usage{
consecutive_group(vec)
}
\arguments{
\item{vec}{the vector}
}
\value{
a vector where the ID depends on whether the value of x is
 sequential. For example, indicate whether a series of years is sequential,
 or if there are say three groups of sequential years.
 Break up the data by groups with consecutive values
}
\description{
Function to add a new column indicating group ID based on sequential data
}
