% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{decision_tables_mp}
\alias{decision_tables_mp}
\title{Make decision table based on MP data}
\usage{
decision_tables_mp(df, xcaption = "Default", xlabel = "tab:default",
  font.size = 9, space.size = 10, placement = "H",
  last_col_header = "TAC", translate = FALSE)
}
\arguments{
\item{df}{data brought in from CSV files found in the herringsr project's data directory}

\item{xcaption}{caption to use for the table}

\item{xlabel}{latex label to use for the table}

\item{font.size}{size of font for table data}

\item{space.size}{space between rows of data in the table}

\item{placement}{latex placement for the table}

\item{last_col_header}{title text for the last column}

\item{translate}{logical. Translate to French if TRUE}
}
\value{
an [xtable::xtable()]
}
\description{
Make decision table based on MP data
}
