% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_catch}
\alias{get_catch}
\alias{get_wa}
\alias{get_pa}
\alias{get_surv_ind}
\title{Summarize catch by region and gear}
\usage{
get_catch(models, models_names, gear, area = 1, group = 1, sex = 0,
  type = 1)

get_wa(models, models_names, gear, area = 1, group = 1, sex = 0)

get_pa(models, models_names, gear, area = 1, group = 1, sex = 0)

get_surv_ind(models, models_names, gear, area = 1, group = 1,
  sex = 0)
}
\arguments{
\item{models}{list of iscam models. Must be a [gfiscamutils::model.lst.class] type}

\item{models_names}{vector of model names that correspond to the input models}

\item{gear}{gear number as it appears in iscam data file}

\item{area}{area number as it appears in iscam data file}

\item{group}{group number as it appears in iscam data file}

\item{sex}{sex number as it appears in iscam data file}

\item{type}{type number as it appears in iscam data file}
}
\value{
a tibble

a tibble

a tibble

a tibble
}
\description{
Summarize catch by region and gear

Summarize weight-at-age data

Summarize proportion-at-age data

Summarize survey index data
}
