% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcr.R
\name{get_hcr_tac_hr}
\alias{get_hcr_tac_hr}
\title{Calculate median of the TAC and HR from the hcr.lst}
\usage{
get_hcr_tac_hr(hcr.lst)
}
\arguments{
\item{hcr.lst}{A list of length-2 vectors, each of which is the TAC and the HR in that order}
}
\value{
a length-2 vector which includes the median TAC and HR for the hcr.lst given
}
\description{
Calculate median of the TAC and HR from the hcr.lst
}
