% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_vars}
\alias{get_vars}
\title{Get model output values for a given region.}
\usage{
get_vars(region, majors = TRUE, french = FALSE, hr_yrs = 10)
}
\arguments{
\item{region}{Character. The short form for the region as defined in the
`major_regions_short` vector`.}

\item{majors}{Logical. If TRUE use major models list (`major_models`), if
FALSE use minor models list (`minor_models`). Default is TRUE.}

\item{french}{Logical. If TRUE, use French (default is FALSE).}
}
\value{
A list of values as required for the report.
}
\description{
Get model output values for a given region.
}
\examples{
hg_vars <- get_vars("HG")
}
