% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcr.R
\name{hcr.hs.slow}
\alias{hcr.hs.slow}
\title{Calculate catch limit based on biomass reference points and reference harvest rate,
 using the last N years for the calculation}
\usage{
hcr.hs.slow(bt, num.end.yrs = 3, ref.hr, lrp, usr, catch.cap, bo = 1)
}
\arguments{
\item{bt}{biomass vector for years}

\item{num.end.yrs}{number of years at the end of the biomass vector to use}

\item{ref.hr}{reference harvest rate}

\item{lrp}{limit reference point}

\item{usr}{upper stock reference point}

\item{catch.cap}{catch cap to use if calculated catch limit is higher}

\item{bo}{initial biomass used in reletive biomass calculation}
}
\value{
catch limit
}
\description{
Calculate catch limit based on biomass reference points and reference harvest rate,
 using the last N years for the calculation
}
