% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcr.R
\name{hcr.min.esc}
\alias{hcr.min.esc}
\title{Calculate catch limit based on a minimum escapement and reference harvest rate}
\usage{
hcr.min.esc(bt, ref.hr, min.esc, abs.esc = FALSE, catch.cap = 0,
  bo = 1)
}
\arguments{
\item{bt}{biomass vector for years}

\item{ref.hr}{reference harvest rate}

\item{min.esc}{minimum escapement to base catch limit on}

\item{abs.esc}{if TRUE, `min.esc` is absolute. If FALSE, it is relative}

\item{catch.cap}{catch cap to use if calculated catch limit is higher}

\item{bo}{initial biomass used in reletive biomass calculation}
}
\value{
catch limit
}
\description{
Calculate catch limit based on a minimum escapement and reference harvest rate
}
