% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcr.R
\name{hcr}
\alias{hcr}
\title{Return the appropriate HCR values for the given 'row' as it appears in the MP data file}
\usage{
hcr(bt, bo, row)
}
\arguments{
\item{bt}{a list of vectors of timeseries biomass values (typically many posteriors)}

\item{bo}{a vector of numeric values (typically many posteriors) for the initial biomass to base
relative biomasses on. This must be the same length as the bt list}

\item{row}{A row of one of the MP tables as it appears in `pbs-assess/herringsr/data/mp-*.csv` files}
}
\value{
A list of length equal to the length of the bt and bo lists, of vectors of length two,
 containing the catch limit (tac) and associated harvest rate (hr)
}
\description{
Return the appropriate HCR values for the given 'row' as it appears in the MP data file
}
