% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{input_data_table}
\alias{input_data_table}
\title{Table showing input data to the herring assessment.}
\usage{
input_data_table(tab, cap = "", translate = FALSE, ...)
}
\arguments{
\item{tab}{data.frame as read in by [readr::read_csv()]}

\item{cap}{caption for table}

\item{translate}{Logical. Translate to french if TRUE}

\item{...}{arguments passed to [csas_table()]}
}
\value{
a [csasdown::csas_table()]
}
\description{
Table showing input data to the herring assessment.
}
