% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{mean_harvest_text}
\alias{mean_harvest_text}
\title{Text for "Application of MPs.." section.}
\usage{
mean_harvest_text(SAR, translate)
}
\arguments{
\item{SAR}{Character. SAR name (e.g., "HG").}

\item{translate}{Logical. Translate to french if TRUE.}
}
\value{
Text.
}
\description{
Text for "Application of MPs.." section.
}
\note{
Requires variables from the `get_vars()` function named for the SAR
  (i.e., `hg_vars <- get_vars("HG", french = french)`).
}
