% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{modify_axes_labels}
\alias{modify_axes_labels}
\title{Add or omit the x and y label and tick mark text to a ggplot.}
\usage{
modify_axes_labels(g, x_label_text = "", y_label_text = "",
  show_x_axis = TRUE, show_y_axis = TRUE, x_axis_label_size = 8,
  x_axis_tick_label_size = 8, y_axis_label_size = 8,
  y_axis_tick_label_size = 8)
}
\arguments{
\item{g}{a ggplot object}

\item{x_label_text}{text for the x label. Only added if `show_x_axis` is `TRUE`}

\item{y_label_text}{text for the x label. Only added if `show_y_axis` is `TRUE`}

\item{show_x_axis}{Logical for whether or not to show the x axis label and tick marks/tick text}

\item{show_y_axis}{Logical for whether or not to show the y axis label and tick marks/tick text}

\item{x_axis_label_size}{size of the font to use for the x axis label}

\item{x_axis_tick_label_size}{size of the font to use for the x axis tick mark labels}

\item{y_axis_label_size}{size of the font to use for the y axis label}

\item{y_axis_tick_label_size}{size of the font to use for the y axis tick mark labels}
}
\value{
the modified ggplot object
}
\description{
Add or omit the x and y label and tick mark text to a ggplot.
}
