% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{newline_format}
\alias{newline_format}
\title{Insert a newline between words so that all single lines will be of length `len` or less.}
\usage{
newline_format(x, len)
}
\arguments{
\item{x}{string to insert newlines into}

\item{len}{length that each line must be}
}
\value{
string with newlines inserted
}
\description{
Insert a newline between words so that all single lines will be of length `len` or less.
}
\examples{
str <- "This is a test of the newline_format function which is used to split text"
newline_format(str, 12)
}
