% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_bh}
\alias{plot_bh}
\title{Plot Beverton-holt}
\usage{
plot_bh(models, regions, translate = FALSE)
}
\arguments{
\item{models}{iscam models list}

\item{regions}{the regions to include}

\item{translate}{Logical; translate labels to french?}
}
\description{
Plot Beverton-holt
}
