% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_biomass_catch}
\alias{plot_biomass_catch}
\title{Plot estimated biommass median and credible interval, projected biomass with credible interval,
catch history, and LRP line with credible interval}
\usage{
plot_biomass_catch(model, catch_df, point_size = 3, errorbar_size = 1,
  line_size = 2, ribbon_alpha = 0.5, lrp_ribbon_alpha = 0.35,
  between_bars = 0.5, refpt_show = "0.3sbo", xlim = NA, annot = NA,
  show_x_axis = TRUE, translate = FALSE)
}
\arguments{
\item{model}{an iscam model object}

\item{catch_df}{a data frame of catch as constructed by [get_catch()]}

\item{point_size}{size of point for projection year median biomass}

\item{errorbar_size}{thickness of errorbar for projection year median biomass}

\item{line_size}{thickness of the median and LRP lines}

\item{ribbon_alpha}{transparency value for the biomass credibility interval ribbon}

\item{lrp_ribbon_alpha}{transparency value for the LRP credibility interval ribbon}

\item{between_bars}{amount of space between catch bars}

\item{refpt_show}{which reference point to show. See `model$mcmccalcs$r.quants`` for choices}

\item{xlim}{x-limits for the plot. Implemented with [ggplot2::coord_cartesian()]}

\item{annot}{a character to place in parentheses in the top left of the plot.
If NA, nothing will appear}

\item{show_x_axis}{Logical}

\item{translate}{Logical. If TRUE, translate to french}
}
\value{
A ggplot object
}
\description{
Plot estimated biommass median and credible interval, projected biomass with credible interval,
catch history, and LRP line with credible interval
}
