% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_biomass_total_biomass}
\alias{plot_biomass_total_biomass}
\title{Plot total and spawning biomass}
\usage{
plot_biomass_total_biomass(
  models,
  reg_names,
  reg_names_short,
  translate = FALSE
)
}
\arguments{
\item{models}{a list of iscam model objects}

\item{reg_names}{region names (must correspond to models)}

\item{reg_names_short}{region names short (must correspond to models)}

\item{translate}{Logical. If TRUE, translate to french}
}
\value{
A ggplot object
}
\description{
Plot total and spawning biomass
}
