% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_catch}
\alias{plot_catch}
\title{Plot catch from a data frames as extracted from iscam data (dat) files}
\usage{
plot_catch(df, xlim = c(1000, 3000), translate = FALSE)
}
\arguments{
\item{df}{a data frame as constructed by [get_catch()]}

\item{xlim}{Limits for the years shown on the plot}

\item{translate}{Logical. If TRUE, translate to French}
}
\value{
A ggplot object
}
\description{
Plot catch from a data frames as extracted from iscam data (dat) files
}
