% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_coastwide_biomass_catch}
\alias{plot_coastwide_biomass_catch}
\title{Plot coastwide spawning biomass and catch}
\usage{
plot_coastwide_biomass_catch(models, catch_df, reg_names, translate = FALSE)
}
\arguments{
\item{models}{a list of iscam model objects}

\item{catch_df}{a data frame as constructed by [get_catch()]}

\item{reg_names}{region names (must correspond to models and catch_df)}

\item{translate}{Logical. If TRUE, translate to french}
}
\value{
A ggplot object
}
\description{
Plot coastwide spawning biomass and catch
}
