% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_harvest_rate}
\alias{plot_harvest_rate}
\title{Effective harvest rate plot}
\usage{
plot_harvest_rate(
  df,
  models,
  regions,
  prop_msy = 0.5,
  show_f_msy = FALSE,
  point_size = 1,
  ribbon_alpha = 0.35,
  ylim = c(0, 1),
  h_line = 0.2,
  translate = FALSE
)
}
\arguments{
\item{df}{a data frame as constructed by [get_catch()]}

\item{models}{a list of iscam model objects}

\item{regions}{a vector of regions names in the order they are to appear in the facets}

\item{prop_msy}{proportion of F_MSY}

\item{show_f_msy}{show horizontal line at prop_msy * F_MSY}

\item{ribbon_alpha}{transparency of the ribbon representing the  credible interval for Ut}

\item{ylim}{Limits to show on the y-axis. Implemented with [ggplot2::coord_cartesian()]}

\item{h_line}{horizontal line value to plot}

\item{translate}{Logical. If TRUE, translate to French}

\item{line_size}{thickness of the mediat Ut line}
}
\value{
a ggplot object
}
\description{
Effective harvest rate plot
}
