% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_hcr}
\alias{plot_hcr}
\title{Plot Harvest control rules as pairs plots}
\usage{
plot_hcr(hcr.lst, sbt.lst, mp = "", region = "", probs = c(0.025,
  0.975), show.medians = TRUE, show.means = TRUE,
  show.x.axes = FALSE, panel.text.size = 4)
}
\arguments{
\item{hcr.lst}{A list of length equal to the length of the sbt list, of vectors of length two,
containing the catch limit (tac) and associated harvest rate (hr)}

\item{sbt.lst}{a list of vectors of timeseries biomass values (typically many posteriors)}

\item{mp}{name of the management procedure to appear on the blank panel}

\item{region}{name for the region to appear on the blank panel}

\item{probs}{vector of length 2 for credible interval of median}

\item{show.medians}{show the median lines and credible intervals}

\item{show.means}{show the mean lines}

\item{show.x.axes}{if TRUE, axes labels, tick marks and tick labels will be shown on the lower plot's x axes}

\item{panel.text.size}{text size for the description panels}
}
\value{
a ggplot object
}
\description{
Plot Harvest control rules as pairs plots
}
