% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_ic}
\alias{plot_ic}
\title{Plot incidental catch from a data frame}
\usage{
plot_ic(df, xlim = c(NA, NA), translate = FALSE)
}
\arguments{
\item{df}{Data frame of incidental catch.}

\item{xlim}{Year limits for the plot (or NA for none).}

\item{translate}{Translate to French?}
}
\value{
A ggplot object.
}
\description{
Plot incidental catch from a data frame
}
