% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_natural_mortality}
\alias{plot_natural_mortality}
\title{Plot natural mortality mcmc median and credibility interval}
\usage{
plot_natural_mortality(model, line_size = 2, ribbon_alpha = 0.5,
  xlim = NA, line_len = 15, annot = NA, show_x_axis = TRUE,
  translate = FALSE)
}
\arguments{
\item{model}{an iscam model}

\item{line_size}{thickness of the median line}

\item{ribbon_alpha}{transparency of the credibility interval ribbon}

\item{xlim}{x-limits for the plot. Implemented with [ggplot2::coord_cartesian()]}

\item{line_len}{length of y-axis label lines (newline will be inserted to make these labels this length)}

\item{annot}{a character to place in parentheses in the top left of the plot.
If NA, nothing will appear}

\item{show_x_axis}{Logical}

\item{translate}{Logical. If TRUE, translate to french}
}
\value{
A ggplot object
}
\description{
Plot natural mortality mcmc median and credibility interval
}
