% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_proj_biomass_density}
\alias{plot_proj_biomass_density}
\title{Plot density of MCMC posterior for projected biomass with quantiles and
 reference point (also with quantiles)}
\usage{
plot_proj_biomass_density(models, regions, yr, tac = 0,
  refpt = "X03B0", line_size = 1, ribbon_alpha = 0.35,
  translate = FALSE)
}
\arguments{
\item{models}{list of iscam model objects}

\item{regions}{a vector of regions names in the order they are to appear in the facets}

\item{yr}{year to plot. Must be in the table `models[[x]]$mcmccalcs$proj.quants``}

\item{tac}{tac to extract data for}

\item{refpt}{the reference point to plot, as found in `models[[x]]$mcmccalcs$proj.quants``}

\item{line_size}{thickness of the mediat Ut line}

\item{ribbon_alpha}{transparency of the ribbon representing the  credible interval for Ut}

\item{translate}{Logical. If TRUE, translate to French}
}
\value{
a ggplot object
}
\description{
Plot density of MCMC posterior for projected biomass with quantiles and
 reference point (also with quantiles)
}
