% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_recruitment}
\alias{plot_recruitment}
\title{Plot recruitment as points and errorbars}
\usage{
plot_recruitment(model, point_size = 2, line_size = 2, xlim = NA,
  annot = NA, show_x_axis = TRUE, translate = FALSE)
}
\arguments{
\item{model}{an iscam model object}

\item{point_size}{Size of points for median recruitment}

\item{line_size}{thickness of errorbars}

\item{xlim}{x-limits for the plot. Implemented with [ggplot2::coord_cartesian()]}

\item{annot}{a character to place in parentheses in the top left of the plot.
If NA, nothing will appear}

\item{show_x_axis}{Logical}

\item{translate}{Logical. If TRUE, translate to french}
}
\value{
A ggplot object
}
\description{
Plot recruitment as points and errorbars
}
