% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_recruitment_devs}
\alias{plot_recruitment_devs}
\title{Plot recruitment deviations as points and errorbars, with a running mean of the median
values}
\usage{
plot_recruitment_devs(model, run_mean_yrs = 3, point_size = 2,
  line_size = 2, errorbar_size = 1, zeroline_size = 1,
  zeroline_type = "dashed", annot = NA, show_x_axis = TRUE,
  translate = FALSE)
}
\arguments{
\item{model}{an iscam model object}

\item{run_mean_yrs}{number of years to set the running mean calculation to}

\item{point_size}{Size of points for median recruitment}

\item{line_size}{thickness of the line for the running mean}

\item{errorbar_size}{thickness of errorbars for recruitment deviations}

\item{zeroline_size}{thickness of guide line at zero deviation}

\item{zeroline_type}{type of  of guide line at zero deviation}

\item{annot}{a character to place in parentheses in the top left of the plot.
If NA, nothing will appear}

\item{show_x_axis}{Logical}

\item{translate}{Logical. If TRUE, translate to french}
}
\value{
A ggplot object
}
\description{
Plot recruitment deviations as points and errorbars, with a running mean of the median
values
}
