% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_spawn_ind}
\alias{plot_spawn_ind}
\title{Plot survey indices from data frames as extracted from iscam data (dat) files}
\usage{
plot_spawn_ind(
  df,
  xlim = c(1000, 3000),
  ylim = NA,
  new_surv_yr = NA,
  new_surv_yr_type = "dashed",
  new_surv_yr_size = 0.25,
  translate = FALSE
)
}
\arguments{
\item{df}{a data frame as constructed by [get_surv_ind()]}

\item{xlim}{limits for the years shown on the plot}

\item{ylim}{limits for the ages shown on the plot}

\item{new_surv_yr}{Year in which the survey type changed. Will be shown as a vertical line}

\item{new_surv_yr_type}{ggplot linetype for new_survey_yr}

\item{new_surv_yr_size}{ggplot line size for new_survey_yr}

\item{translate}{Logical. If TRUE, translate to French}
}
\value{
A ggplot object
}
\description{
Plot survey indices from data frames as extracted from iscam data (dat) files
}
