% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_spawn_section}
\alias{plot_spawn_section}
\title{Plot spawn index or scaled abundance by section}
\usage{
plot_spawn_section(
  model,
  data_file,
  sections = NA,
  yr_range = NA,
  scale = TRUE,
  ncol = 2,
  translate = FALSE
)
}
\arguments{
\item{model}{iscam model object (can be NA if scale is FALSE).}

\item{data_file}{Chracter. Path and name of data file (CSV) with spawn by
section and year.}

\item{sections}{Vector of sections to include, or NA to include all.}

\item{yr_range}{Start and end year, or NA to include all the data.}

\item{scale}{Logical. Scale the spawn index using q?}

\item{translate}{Logical. Translate to French?}
}
\description{
Plot spawn index or scaled abundance by section
}
