% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{prod_period_table}
\alias{prod_period_table}
\title{Table showing productive period and related information}
\usage{
prod_period_table(dat, cap = "", translate = FALSE, ...)
}
\arguments{
\item{dat}{list with SARs, years, and proportion}

\item{cap}{caption for table}

\item{translate}{Logical. Translate to french if TRUE}

\item{...}{arguments passed to [csas_table()]}
}
\value{
a [csasdown::csas_table()]
}
\description{
Table showing productive period and related information
}
