% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{sok_harvest_table}
\alias{sok_harvest_table}
\title{Table for the Spawn on kelp harvest for herring}
\usage{
sok_harvest_table(tab, by_vec, first_yr, cap = "", translate = FALSE,
  ...)
}
\arguments{
\item{tab}{data.frame as read in by [readr::read_csv()]}

\item{by_vec}{a vector of names, which correspond to the names of the columns in tab}

\item{first_yr}{first year to show in table}

\item{cap}{caption for table}

\item{translate}{Logical. Translate to french if TRUE}

\item{...}{arguments passed to [csas_table()]}
}
\value{
a [csasdown::csas_table()]
}
\description{
Table for the Spawn on kelp harvest for herring
}
