% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{stock_status_text}
\alias{stock_status_text}
\title{Text for "Stock status update" section.}
\usage{
stock_status_text(SAR, this_yr, p_lrp = 0.3, translate)
}
\arguments{
\item{SAR}{Character. SAR name (e.g., "HG").}

\item{this_yr}{Numeric. Usually `assess_yr`.}

\item{p_lrp}{Numeric. Proportion of SB_0 that is the LRP (default is 0.3).}

\item{translate}{Logical. Translate to french if TRUE.}
}
\value{
Text.
}
\description{
Text for "Stock status update" section.
}
\note{
Requires variables from the `get_vars()` function named for the SAR
  (i.e., `hg_vars <- get_vars("HG", french = french)`).
}
