% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{total_landed_catch_table}
\alias{total_landed_catch_table}
\title{Table showing the total landed catch by area for herring}
\usage{
total_landed_catch_table(tab, by_vec, first_yr, cap = "",
  translate = FALSE, ...)
}
\arguments{
\item{tab}{data.frame as returned from [gfiscamutils::get_catch()]}

\item{by_vec}{a vector of names, which correspond to the names of the columns in tab}

\item{first_yr}{first year to show in table}

\item{cap}{caption for table}

\item{translate}{Logical. Translate to french if TRUE}

\item{...}{arguments passed to [csas_table()]}
}
\value{
a [csasdown::csas_table()]
}
\description{
Table showing the total landed catch by area for herring
}
