% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{decision_tables_mp}
\alias{decision_tables_mp}
\title{Make decision table based on MP data}
\usage{
decision_tables_mp(df, xcaption = "Default", xlabel = "tab:default",
  font.size = 9, space.size = 10, placement = "H",
  last_col_header = "TAC", perc_dec_pts = 0, dec_pts = 2,
  col_align = "ccc|r|r|c|c|r|c", translate = FALSE)
}
\arguments{
\item{df}{data brought in from CSV files found in the herringsr project's data directory}

\item{xcaption}{caption to use for the table}

\item{xlabel}{latex label to use for the table}

\item{font.size}{size of font for table data}

\item{space.size}{space between rows of data in the table}

\item{placement}{latex placement for the table}

\item{last_col_header}{title text for the last column}

\item{perc_dec_pts}{number of decimal points to show for percentage columns}

\item{dec_pts}{number of decimal points to show for numerical non-percentage columns}

\item{col_align}{string for alignment of columns. c=center, r=right, l=left, |=place vertical bar between column}

\item{translate}{logical. Translate to French if TRUE}
}
\value{
an [xtable::xtable()]
}
\description{
Make decision table based on MP data
}
