% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{decision_tables_mp}
\alias{decision_tables_mp}
\title{Make decision table based on MP data}
\usage{
decision_tables_mp(
  df,
  xcaption = "Default",
  xlabel = "tab:default",
  p_sb_ave_prod = "",
  font.size = 11,
  space.size = 15,
  placement = "ht",
  perc_dec_pts = 0,
  dec_pts = 2,
  inc_mps = NA,
  hide_tac_hr_below_targ = FALSE,
  translate = FALSE
)
}
\arguments{
\item{df}{data brought in from CSV files found in the herringsr project's data directory}

\item{xcaption}{caption to use for the table}

\item{xlabel}{latex label to use for the table}

\item{p_sb_ave_prod}{Proportion of average SB during the productive period}

\item{font.size}{size of font for table data}

\item{space.size}{space between rows of data in the table}

\item{placement}{latex placement for the table}

\item{perc_dec_pts}{number of decimal points to show for percentage columns}

\item{dec_pts}{number of decimal points to show for numerical non-percentage columns}

\item{inc_mps}{include these MP numbers in the table. If NA, all will be included}

\item{translate}{logical. Translate to French if TRUE}

\item{col_align}{string for alignment of columns. c=center, r=right, l=left, |=place vertical bar between column}
}
\value{
an [xtable::xtable()]
}
\description{
Make decision table based on MP data
}
