% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcr.R
\name{hcr.hs}
\alias{hcr.hs}
\title{Calculate catch limit based on biomass reference points and reference harvest rate}
\usage{
hcr.hs(bt, ref.hr, lrp, usr, catch.cap, bo = 1)
}
\arguments{
\item{bt}{biomass vector for years}

\item{ref.hr}{reference harvest rate}

\item{lrp}{limit reference point}

\item{usr}{upper stock reference point}

\item{catch.cap}{catch cap to use if calculated catch limit is higher}

\item{bo}{initial biomass used in reletive biomass calculation}
}
\value{
catch limit
}
\description{
Calculate catch limit based on biomass reference points and reference harvest rate
}
