% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcr.R
\name{hcr}
\alias{hcr}
\title{Return the appropriate HCR values for the given 'row' as it appears in the MP data file}
\usage{
hcr(bt, bo, row)
}
\arguments{
\item{bt}{a list of vectors of timeseries biomass values}

\item{bo}{a vector of numeric values for the initial biomass to base relative biomasses on. This must be the
same length as the bt list}

\item{row}{A row of one of the MP tables as it appears in `pbs-assess/herringsr/data/mp-*.csv` files}
}
\value{
The original row with the catch limit (catch) and associated harvest rate (hr) populated
}
\description{
Return the appropriate HCR values for the given 'row' as it appears in the MP data file
}
