% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_biomass_catch}
\alias{plot_biomass_catch}
\title{Plot estimated biommass median and credible interval, projected biomass with credible interval,
catch history, and LRP line with credible interval}
\usage{
plot_biomass_catch(
  model,
  catch_df,
  point_size = 1,
  errorbar_size = 1,
  line_size = 0.75,
  ribbon_alpha = 0.5,
  lrp_ribbon_alpha = 0.35,
  show_lrp_ribbon = TRUE,
  show_catch = TRUE,
  between_bars = 0.75,
  refpt_show = "0.3sbo",
  show_prod_usr = TRUE,
  prod_yrs = 1990:1999,
  prop_prod = 1,
  show_prod_yrs = TRUE,
  show_sbo_usr = FALSE,
  prop_sbo = 0.5,
  show_blt_usr = FALSE,
  prop_blt = 1,
  col_catch = TRUE,
  xlim = NA,
  show_x_axis = TRUE,
  show_y_axis = TRUE,
  x_axis_label_size = 8,
  x_axis_tick_label_size = 8,
  y_axis_label_size = 8,
  y_axis_tick_label_size = 8,
  x_axis_label_newline_length = 50,
  y_axis_label_newline_length = 40,
  x_axis_position = "bottom",
  y_axis_position = "left",
  annot = "d",
  label_points = FALSE,
  label_round = 1,
  translate = FALSE
)
}
\arguments{
\item{model}{an iscam model object}

\item{catch_df}{a data frame of catch as constructed by [get_catch()]}

\item{point_size}{size of point for projection year median biomass}

\item{errorbar_size}{thickness of errorbar for projection year median biomass}

\item{line_size}{thickness of the median and LRP lines}

\item{ribbon_alpha}{transparency value for the biomass credibility interval ribbon}

\item{lrp_ribbon_alpha}{transparency value for the LRP credibility interval ribbon}

\item{between_bars}{amount of space between catch bars}

\item{refpt_show}{which reference point to show. See `model$mcmccalcs$r.quants`` for choices}

\item{show_prod_usr}{Logical. Show the productive period USR. Default TRUE.}

\item{prod_yrs}{Numeric vector. Productive period to calculate the USR.
Default 1990:1999.}

\item{prop_prod}{Numeric. Proportion of productive period for USR. Default
1.0.}

\item{show_prod_yrs}{Logical. Show vertical band for productive period.
Default TRUE.}

\item{show_sbo_usr}{Logical. Show SB_0. Default TRUE.}

\item{prop_sbo}{Numeric. Proportion of SB_0 for USR. Default 0.6.}

\item{show_blt_usr}{Logical. Show average long-term biomass USR. Default TRUE.}

\item{prop_blt}{Numeric. Proportion of average long-term biomass for USR. Default 1.0.}

\item{col_catch}{Logical. Colour catch bars, or just grey. Default TRUE.}

\item{xlim}{x-limits for the plot. Implemented with [ggplot2::coord_cartesian()]}

\item{show_x_axis}{see [modify_axes_labels()]}

\item{show_y_axis}{see [modify_axes_labels()]}

\item{x_axis_label_size}{see [modify_axes_labels()]}

\item{x_axis_tick_label_size}{see [modify_axes_labels()]}

\item{y_axis_label_size}{see [modify_axes_labels()]}

\item{y_axis_tick_label_size}{see [modify_axes_labels()]}

\item{x_axis_label_newline_length}{see [newline_format()]}

\item{y_axis_label_newline_length}{see [newline_format()]}

\item{annot}{a character to place in parentheses in the top left of the plot.
If NA, nothing will appear}

\item{translate}{Logical. If TRUE, translate to french}
}
\value{
A ggplot object
}
\description{
Plot estimated biommass median and credible interval, projected biomass with credible interval,
catch history, and LRP line with credible interval
}
