% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_biomass_phase}
\alias{plot_biomass_phase}
\title{Plot production using a phase plot with a path through time. Includes the reference point and its credible interval}
\usage{
plot_biomass_phase(model, catch_df, new_surv_yr = NA, point_size = 3,
  line_size = 2, text_size = 2, zeroline_size = 1,
  zeroline_type = "dashed", lrp_ribbon_alpha = 0.35,
  refpt_show = "0.3sbo", line_len = 15, annot = NA,
  show_x_axis = TRUE, translate = FALSE)
}
\arguments{
\item{model}{an iscam model object}

\item{catch_df}{a data frame of catch as constructed by [get_catch()]}

\item{new_surv_yr}{year in which the survey changed from surface to dive}

\item{point_size}{size for points for years}

\item{line_size}{thickness of the path line}

\item{text_size}{size for year marker text labels}

\item{zeroline_size}{thickness of the line across zero}

\item{zeroline_type}{type of the line across zer}

\item{lrp_ribbon_alpha}{transparency of the reference point credible interval ribbon}

\item{refpt_show}{which reference point to show. See `model$mcmccalcs$r.quants`` for choices}

\item{line_len}{length of y-axis label lines (newline will be inserted to make these labels this length)}

\item{annot}{a character to place in parentheses in the top left of the plot.
If NA, nothing will appear}

\item{show_x_axis}{Logical}

\item{translate}{Logical. If TRUE, translate to french}
}
\value{
A ggplot object
}
\description{
Plot production using a phase plot with a path through time. Includes the reference point and its credible interval
}
