% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_hcr}
\alias{plot_hcr}
\title{Plot Harvest control rules as pairs plots}
\usage{
plot_hcr(
  hcr.lst,
  sbt.lst,
  sbo.lst,
  mp = "",
  region = "",
  probs = c(0.025, 0.975),
  depletion = TRUE,
  point.size = 0.3,
  point.alpha = 0.4,
  line.width = 0.5,
  show.medians = TRUE,
  show.means = TRUE,
  show.shaded = FALSE,
  show.x.axes = TRUE,
  axis.text.size = 7,
  panel.text.size = 3
)
}
\arguments{
\item{hcr.lst}{A list of length equal to the length of the sbt list, of vectors of length two,
containing the catch limit (tac) and associated harvest rate (hr)}

\item{sbt.lst}{a list of vectors of timeseries biomass values (typically many posteriors)}

\item{sbo.lst}{a list of vectors of timeseries unfish biomass values (typically many posteriors)}

\item{mp}{name of the management procedure to appear on the blank panel}

\item{region}{name for the region to appear on the blank panel}

\item{probs}{vector of length 2 for credible interval of median}

\item{depletion}{if TRUE, show depletion (SB_t/SB_0); if FALSE show SB_t}

\item{point.size}{size of points}

\item{point.alpha}{transparency of points}

\item{line.width}{width ot thickness of all lines}

\item{show.medians}{show the median lines and credible intervals}

\item{show.means}{show the mean lines}

\item{show.x.axes}{if TRUE, axes labels, tick marks and tick labels will be shown on the lower plot's x axes}

\item{axis.text.size}{size of text for axis labels}

\item{panel.text.size}{text size for the description panels}
}
\value{
a ggplot object
}
\description{
Plot Harvest control rules as pairs plots
}
