% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_recruitment}
\alias{plot_recruitment}
\title{Plot recruitment as points and errorbars}
\usage{
plot_recruitment(
  model,
  point_size = 2,
  line_size = 2,
  xlim = NA,
  show_x_axis = TRUE,
  show_y_axis = TRUE,
  x_axis_label_size = 8,
  x_axis_tick_label_size = 8,
  y_axis_label_size = 8,
  y_axis_tick_label_size = 8,
  x_axis_label_newline_length = 30,
  y_axis_label_newline_length = 20,
  annot = NA,
  translate = FALSE
)
}
\arguments{
\item{model}{an iscam model object}

\item{point_size}{Size of points for median recruitment}

\item{line_size}{thickness of errorbars}

\item{xlim}{x-limits for the plot. Implemented with [ggplot2::coord_cartesian()]}

\item{show_x_axis}{see [modify_axes_labels()]}

\item{show_y_axis}{see [modify_axes_labels()]}

\item{x_axis_label_size}{see [modify_axes_labels()]}

\item{x_axis_tick_label_size}{see [modify_axes_labels()]}

\item{y_axis_label_size}{see [modify_axes_labels()]}

\item{y_axis_tick_label_size}{see [modify_axes_labels()]}

\item{x_axis_label_newline_length}{see [newline_format()]}

\item{y_axis_label_newline_length}{see [newline_format()]}

\item{annot}{a character to place in parentheses in the top left of the plot.
If NA, nothing will appear}

\item{translate}{Logical. If TRUE, translate to french}
}
\value{
A ggplot object
}
\description{
Plot recruitment as points and errorbars
}
