% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_recruitment}
\alias{plot_recruitment}
\title{Plot recruitment as points and errorbars}
\usage{
plot_recruitment(
  model,
  point_size = 1,
  line_size = 1,
  prod_yrs = 1990:1999,
  show_prod_yrs = FALSE,
  xlim = NA,
  show_x_axis = FALSE,
  show_y_axis = TRUE,
  x_axis_label_size = 8,
  x_axis_tick_label_size = 8,
  y_axis_label_size = 8,
  y_axis_tick_label_size = 8,
  x_axis_label_newline_length = 50,
  y_axis_label_newline_length = 40,
  annot = "c",
  show_r0 = TRUE,
  line_size_r0 = 0.5,
  r0_ribbon_alpha = 0.35,
  translate = FALSE
)
}
\arguments{
\item{model}{an iscam model object}

\item{point_size}{Size of points for median recruitment}

\item{line_size}{thickness of errorbars}

\item{prod_yrs}{Numeric vector. Productive period to calculate the USR.
Default 1990:1999.}

\item{show_prod_yrs}{Logical. Show vertical band for productive period.
Default TRUE.}

\item{xlim}{x-limits for the plot. Implemented with
[ggplot2::coord_cartesian()]}

\item{show_x_axis}{see [modify_axes_labels()]}

\item{show_y_axis}{see [modify_axes_labels()]}

\item{x_axis_label_size}{see [modify_axes_labels()]}

\item{x_axis_tick_label_size}{see [modify_axes_labels()]}

\item{y_axis_label_size}{see [modify_axes_labels()]}

\item{y_axis_tick_label_size}{see [modify_axes_labels()]}

\item{x_axis_label_newline_length}{see [newline_format()]}

\item{y_axis_label_newline_length}{see [newline_format()]}

\item{annot}{a character to place in parentheses in the top left of the plot.
If NA, nothing will appear}

\item{show_r0}{Add horizontal like for unfished recruitment R_0, and shaded
rectangle for CI.}

\item{line_size_r0}{Thickness for horizontal line}

\item{r0_ribbon_alpha}{Alpha for R_0 rectangle (uncertainty).}

\item{translate}{Logical. If TRUE, translate to french}
}
\value{
A ggplot object
}
\description{
Plot recruitment as points and errorbars
}
