% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_recruitment_devs}
\alias{plot_recruitment_devs}
\title{Plot recruitment deviations as points and errorbars, with a running mean of the median
values}
\usage{
plot_recruitment_devs(
  model,
  run_mean_yrs = 3,
  point_size = 1,
  line_size = 0.5,
  errorbar_size = 0.5,
  zeroline_size = 0.5,
  zeroline_type = "dashed",
  prod_yrs = 1990:1999,
  show_prod_yrs = FALSE,
  xlim = NA,
  show_x_axis = TRUE,
  show_y_axis = TRUE,
  x_axis_label_size = 8,
  x_axis_tick_label_size = 8,
  y_axis_label_size = 8,
  y_axis_tick_label_size = 8,
  x_axis_label_newline_length = 50,
  y_axis_label_newline_length = 40,
  annot = "e",
  translate = FALSE
)
}
\arguments{
\item{model}{an iscam model object}

\item{run_mean_yrs}{number of years to set the running mean calculation to}

\item{point_size}{Size of points for median recruitment}

\item{line_size}{thickness of the line for the running mean}

\item{errorbar_size}{thickness of errorbars for recruitment deviations}

\item{zeroline_size}{thickness of guide line at zero deviation}

\item{zeroline_type}{type of  of guide line at zero deviation}

\item{prod_yrs}{Numeric vector. Productive period to calculate the USR.
Default 1990:1999.}

\item{show_prod_yrs}{Logical. Show vertical band for productive period.
Default TRUE.}

\item{xlim}{x-limits for the plot. Implemented with [ggplot2::coord_cartesian()]}

\item{show_x_axis}{see [modify_axes_labels()]}

\item{show_y_axis}{see [modify_axes_labels()]}

\item{x_axis_label_size}{see [modify_axes_labels()]}

\item{x_axis_tick_label_size}{see [modify_axes_labels()]}

\item{y_axis_label_size}{see [modify_axes_labels()]}

\item{y_axis_tick_label_size}{see [modify_axes_labels()]}

\item{x_axis_label_newline_length}{see [newline_format()]}

\item{y_axis_label_newline_length}{see [newline_format()]}

\item{annot}{a character to place in parentheses in the top left of the plot.
If NA, nothing will appear}

\item{translate}{Logical. If TRUE, translate to french}
}
\value{
A ggplot object
}
\description{
Plot recruitment deviations as points and errorbars, with a running mean of the median
values
}
