% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_wa}
\alias{plot_wa}
\title{Plot weight-at-age time series from a data frames as extracted from iscam data (dat) files}
\usage{
plot_wa(
  df,
  circle_age = 3,
  xlim = c(1000, 3000),
  ylim = c(0, NA),
  n_roll = 5,
  translate = FALSE
)
}
\arguments{
\item{df}{a data frame as constructed by [get_wa()]}

\item{circle_age}{the age for which to add circles to plot}

\item{xlim}{Limits for the years shown on the plot}

\item{ylim}{limits for the weights shown on the plot}

\item{n_roll}{Number of years to calculate the rolling mean (window)}

\item{translate}{Logical. If TRUE, translate to French}
}
\value{
A ggplot object
}
\description{
Plot weight-at-age time series from a data frames as extracted from iscam data (dat) files
}
